+++ src/serializers/interfaces/mocks.ts
@@ -1,0 +1,8 @@
+import { Mocked, vi } from "vitest";
+
+import { IHttpErrorSerializer } from "./IHttpErrorSerializer.js";
+
+export const mockHttpErrorSerializer: Mocked<IHttpErrorSerializer> = {
+  serializerError: vi.fn(),
+  serializeResponse: vi.fn(),
+};